+++ scripts/vespa_local/pom.xml
@@ -16,34 +16,11 @@
             <version>5.8.2</version>
             <scope>test</scope>
         </dependency>
-        <!--
-        <dependency> 
+        <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
-            <version>2.0.13</version>
+            <scope>provided</scope>
         </dependency>
-
-        <dependency>              
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-api</artifactId>
-            <version>2.23.1</version>
-        </dependency>
-        -->
-        <!--
-        <dependency>              
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-core</artifactId>
-            <version>2.23.1</version>
-        </dependency>
-        -->
-        <!--
-        <dependency>
-            <groupId>com.google.errorprone</groupId>
-            <artifactId>error_prone_annotations</artifactId>
-            <version>2.28.0</version>
-            <scope>compile</scope>
-        </dependency>
-        -->
     </dependencies>
     <build>
         <plugins>
